backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncpg>=0.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
websockets>=11.0.0
python-socketio>=5.8.0
redis>=4.6.0
aiofiles>=23.1.0
tenacity>=8.2.0
prometheus-client>=0.17.0
structlog>=23.1.0